mysql-ai-agent/requirements.txt
pymysql
pandas
langchain
langchain-core
langchain-openai
python-dotenv
pydantic
pytest